Move contact form & widget to separate files
