+++ tracpro/contacts/fields.py
@@ -1,0 +1,44 @@
+from django import forms
+from django.utils.safestring import mark_safe
+from django.utils.translation import ugettext_lazy as _
+
+
+URN_SCHEME_TEL = 'tel'
+URN_SCHEME_TWITTER = 'twitter'
+URN_SCHEME_CHOICES = (
+    (URN_SCHEME_TEL, _("Phone")),
+    (URN_SCHEME_TWITTER, _("Twitter")),
+)
+
+
+class URNWidget(forms.widgets.MultiWidget):
+
+    def __init__(self, *args, **kwargs):
+        scheme_choices = kwargs.pop('scheme_choices')
+        widgets = (forms.Select(choices=scheme_choices),
+                   forms.TextInput(attrs={'maxlength': 32}))
+        super(URNWidget, self).__init__(widgets, *args, **kwargs)
+
+    def decompress(self, value):
+        if value:
+            return value.split(':', 1)
+        else:
+            return URN_SCHEME_TEL, ''
+
+    def render(self, name, value, attrs=None):
+        output = ['<div class="urn-widget">',
+                  super(URNWidget, self).render(name, value, attrs),
+                  '</div>']
+        return mark_safe(''.join(output))
+
+
+class URNField(forms.fields.MultiValueField):
+
+    def __init__(self, *args, **kwargs):
+        fields = (forms.ChoiceField(choices=URN_SCHEME_CHOICES),
+                  forms.CharField(max_length=32))
+        super(URNField, self).__init__(fields, *args, **kwargs)
+        self.widget = URNWidget(scheme_choices=URN_SCHEME_CHOICES)
+
+    def compress(self, values):
+        return '%s:%s' % (values[0], values[1])
