+++ tracpro/contacts/views.py
@@ -8,95 +8,20 @@ from dash.orgs.views import OrgPermsMixin, OrgObjPermsMixin
 from dash.utils import get_obj_cacheable
 from dash.utils.sync import ChangeType
 
-from django import forms
 from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect, JsonResponse
-from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
 from smartmin.users.views import (
     SmartListView, SmartCreateView, SmartReadView, SmartUpdateView,
     SmartDeleteView, SmartCRUDL)
 
-from tracpro.groups.models import Region, Group
 from tracpro.polls.models import PollRun, RESPONSE_COMPLETE
 
+from .fields import URN_SCHEME_CHOICES
+from .forms import ContactForm
 from .models import Contact
-
-
-URN_SCHEME_TEL = 'tel'
-URN_SCHEME_TWITTER = 'twitter'
-URN_SCHEME_CHOICES = (
-    (URN_SCHEME_TEL, _("Phone")),
-    (URN_SCHEME_TWITTER, _("Twitter")),
-)
-
-
-class URNField(forms.fields.MultiValueField):
-
-    def __init__(self, *args, **kwargs):
-        fields = (forms.ChoiceField(choices=URN_SCHEME_CHOICES),
-                  forms.CharField(max_length=32))
-        super(URNField, self).__init__(fields, *args, **kwargs)
-        self.widget = URNWidget(scheme_choices=URN_SCHEME_CHOICES)
-
-    def compress(self, values):
-        return '%s:%s' % (values[0], values[1])
-
-
-class URNWidget(forms.widgets.MultiWidget):
-
-    def __init__(self, *args, **kwargs):
-        scheme_choices = kwargs.pop('scheme_choices')
-        widgets = (forms.Select(choices=scheme_choices),
-                   forms.TextInput(attrs={'maxlength': 32}))
-        super(URNWidget, self).__init__(widgets, *args, **kwargs)
-
-    def decompress(self, value):
-        if value:
-            return value.split(':', 1)
-        else:
-            return URN_SCHEME_TEL, ''
-
-    def render(self, name, value, attrs=None):
-        output = ['<div class="urn-widget">',
-                  super(URNWidget, self).render(name, value, attrs),
-                  '</div>']
-        return mark_safe(''.join(output))
-
-
-class ContactForm(forms.ModelForm):
-    name = forms.CharField(max_length=128, label=_("Full name"))
-    urn = URNField(
-        label=_("Phone/Twitter"),
-        help_text=_("Phone number or Twitter handle of this contact."))
-    region = forms.ModelChoiceField(
-        label=_("Region"),
-        queryset=Region.objects.filter(pk=-1),
-        help_text=_("Region where this contact lives."))
-    group = forms.ModelChoiceField(
-        label=_("Reporter Group"),
-        queryset=Group.objects.filter(pk=-1),
-        help_text=_("Reporter Group to which this contact belongs."))
-    facility_code = forms.CharField(
-        max_length=16, label=_("Facility Code"), required=False)
-    language = forms.CharField(
-        label=_("Language"), required=False,
-        widget=forms.TextInput(attrs={'class': 'language-field'}))
-
-    class Meta:
-        model = Contact
-        fields = forms.ALL_FIELDS
-
-    def __init__(self, *args, **kwargs):
-        self.user = kwargs.pop('user')
-        org = self.user.get_org()
-
-        super(ContactForm, self).__init__(*args, **kwargs)
-
-        self.fields['region'].queryset = self.user.get_regions(org).order_by('name')
-        self.fields['group'].queryset = Group.get_all(org).order_by('name')
 
 
 class ContactFormMixin(object):
